zombienet_sdk
Struct Network

pub struct Network<T>
where
    T: FileSystem,
{ /* private fields */ }

Implementations

impl<T> Network<T>
where T: FileSystem,

pub fn ns_name(&self) -> String

pub fn base_dir(&self) -> Option<&str>

pub fn relaychain(&self) -> &Relaychain

pub async fn destroy(self) -> Result<(), ProviderError>

pub async fn add_node( &mut self, name: impl Into<String>, options: AddNodeOptions, ) -> Result<(), Error>

Add a node to the relaychain

NOTE: name must be unique in the whole network. The new node is added to the running network instance.

Example:

let mut network = orchestrator.spawn(config).await?;

// Create the options to add the new node
let opts = AddNodeOptions {
    rpc_port: Some(9444),
    is_validator: true,
    ..Default::default()
};

network.add_node("new-node", opts).await?;

pub async fn add_collator( &mut self, name: impl Into<String>, options: AddCollatorOptions, para_id: u32, ) -> Result<(), Error>

Add a new collator to a parachain

NOTE: name must be unique in the whole network.

Example:

let mut network = orchestrator.spawn(config).await?;

let col_opts = AddCollatorOptions {
    command: Some("polkadot-parachain".try_into()?),
    ..Default::default()
};

network.add_collator("new-col-1", col_opts, 100).await?;

pub fn para_config_builder(&self) -> ParachainConfigBuilder<Initial, Running>

Get a parachain config builder from a running network

This allow you to build a new parachain config to be deployed into the running network.

pub async fn add_parachain( &mut self, para_config: &ParachainConfig, custom_relaychain_spec: Option<PathBuf>, custom_parchain_fs_prefix: Option<String>, ) -> Result<(), Error>

Add a new parachain to the running network

Arguments

para_config - Parachain configuration to deploy

custom_relaychain_spec - Optional path to a custom relaychain spec to use

custom_parchain_fs_prefix - Optional prefix to use when artifacts are created

Example:

let mut network = orchestrator.spawn(config).await?;
let para_config = network
    .para_config_builder()
    .with_id(100)
    .with_default_command("polkadot-parachain")
    .with_collator(|c| c.with_name("col-100-1"))
    .build()
    .map_err(|_e| anyhow!("Building config"))?;

network.add_parachain(&para_config, None, None).await?;

pub fn get_node(&self, name: impl Into<String>) -> Result<&NetworkNode, Error>

pub fn get_node_mut( &mut self, name: impl Into<String>, ) -> Result<&mut NetworkNode, Error>

pub fn nodes(&self) -> Vec<&NetworkNode>

pub async fn detach(&self)

pub fn name(&self) -> &str

pub fn parachain(&self, para_id: u32) -> Option<&Parachain>

pub fn parachains(&self) -> Vec<&Parachain>