zombienet_sdk
Struct NetworkNode

pub struct NetworkNode { /* private fields */ }

Implementations

impl NetworkNode

pub fn name(&self) -> &str

pub fn args(&self) -> Vec<&str>

pub fn spec(&self) -> &NodeSpec

pub fn ws_uri(&self) -> &str

pub async fn rpc(&self) -> Result<RpcClient, Error>

Get the rpc client for the node

pub async fn client<Config>(&self) -> Result<OnlineClient<Config>, Error>
where Config: Config,

👎Deprecated: Use wait_client instead.

Get the online client for the node

pub async fn try_client<Config>(&self) -> Result<OnlineClient<Config>, Error>
where Config: Config,

Try to connect to the node.

Most of the time you only want to use NetworkNode::wait_client that waits for the node to appear before it connects to it. This function directly tries to connect to the node and returns an error if the node is not yet available at that point in time.

Returns a [OnlineClient] on success.

pub async fn wait_client<Config>(&self) -> Result<OnlineClient<Config>, Error>
where Config: Config,

Wait until get the online client for the node

pub async fn wait_client_with_timeout<Config>( &self, timeout_secs: impl Into<u64>, ) -> Result<OnlineClient<Config>, Error>
where Config: Config,

Wait until get the online client for the node with a defined timeout

pub async fn pause(&self) -> Result<(), Error>

Pause the node, this is implemented by pausing the actual process (e.g polkadot) with sending SIGSTOP signal

pub async fn resume(&self) -> Result<(), Error>

Resume the node, this is implemented by resuming the actual process (e.g polkadot) with sending SIGCONT signal

pub async fn restart(&self, after: Option<Duration>) -> Result<(), Error>

Restart the node using the same cmd, args and env (and same isolated dir)

pub async fn reports( &self, metric_name: impl Into<String>, ) -> Result<f64, Error>

Get metric value ‘by name’ from Prometheus (exposed by the node) metric name can be: with prefix (e.g: ‘polkadot_’) with chain attribute (e.g: ‘chain=rococo-local’) without prefix and/or without chain attribute

pub async fn assert( &self, metric_name: impl Into<String>, value: impl Into<f64>, ) -> Result<bool, Error>

Assert on a metric value ‘by name’ from Prometheus (exposed by the node) metric name can be: with prefix (e.g: ‘polkadot_’) with chain attribute (e.g: ‘chain=rococo-local’) without prefix and/or without chain attribute

We first try to assert on the value using the cached metrics and if not meet the criteria we reload the cache and check again

pub async fn assert_with( &self, metric_name: impl Into<String>, predicate: impl Fn(f64) -> bool, ) -> Result<bool, Error>

Assert on a metric value using a given predicate. See NetworkNode::reports description for details on metric name.

pub async fn wait_metric( &self, metric_name: impl Into<String>, predicate: impl Fn(f64) -> bool, ) -> Result<(), Error>

Wait until a metric value pass the predicate

pub async fn wait_metric_with_timeout( &self, metric_name: impl Into<String>, predicate: impl Fn(f64) -> bool, timeout_secs: impl Into<u64>, ) -> Result<(), Error>

Wait until a metric value pass the predicate with a timeout (secs)

pub async fn logs(&self) -> Result<String, Error>

Get the logs of the node TODO: do we need the since param, maybe we could be handy later for loop filtering

pub async fn wait_log_line_count( &self, pattern: impl Into<String>, is_glob: bool, count: usize, ) -> Result<(), Error>

Wait until a the number of matching log lines is reach

pub async fn wait_log_line_count_with_timeout( &self, substring: impl Into<String>, is_glob: bool, count: usize, timeout_secs: impl Into<u64>, ) -> Result<(), Error>

Wait until a the number of matching log lines is reach with timeout (secs)

pub async fn pjs( &self, code: impl AsRef<str>, args: Vec<Value>, user_types: Option<Value>, ) -> Result<Result<Value, String>, Error>

Execute js/ts code inside [pjs_rs] custom runtime.

The code will be run in a wrapper similar to the javascript developer tab of polkadot.js apps. The returning value is represented as PjsResult enum, to allow to communicate that the execution was successful but the returning value can be deserialized as serde_json::Value.

pub async fn pjs_file( &self, file: impl AsRef<Path>, args: Vec<Value>, user_types: Option<Value>, ) -> Result<Result<Value, String>, Error>

Execute js/ts file inside [pjs_rs] custom runtime.

The content of the file will be run in a wrapper similar to the javascript developer tab of polkadot.js apps. The returning value is represented as PjsResult enum, to allow to communicate that the execution was successful but the returning value can be deserialized as serde_json::Value.