Crate zombienet_sdk

Structs

AddCollatorOptions

AddNodeOptions

LocalFileSystem

Network

NetworkConfig

A network configuration, composed of a relaychain, parachains and HRMP channels.

NetworkConfigBuilder

A network configuration builder, used to build a NetworkConfig declaratively with fields validation.

NetworkNode

Orchestrator

Enums

OrchestratorError

RegistrationStrategy

The registration strategy that will be used for the parachain.

Constants

PROVIDERS

Traits

NetworkConfigExt

Type Aliases

PjsResult

pjs-rs success Result type