Crate zombienet_sdk

Re-exports

pub use orchestrator::sc_chain_spec;

pub use subxt;

pub use subxt_signer;

Modules

environment

Helpers functions to get configuration (e.g. Provider and images) from the env vars

tx_helper

Structs

AddCollatorOptions

AddNodeOptions

AttachToLiveNetwork

GlobalSettings

Global settings applied to an entire network.

GlobalSettingsBuilder

A global settings builder, used to build GlobalSettings declaratively with fields validation.

LocalFileSystem

Network

NetworkConfig

A network configuration, composed of a relaychain, parachains and HRMP channels.

NetworkConfigBuilder

A network configuration builder, used to build a NetworkConfig declaratively with fields validation.

NetworkNode

Orchestrator

Enums

OrchestratorError

RegistrationStrategy

The registration strategy that will be used for the parachain.

WithRelaychain

Constants

PROVIDERS

Traits

AttachToLive

NetworkConfigExt